+++ consensus/pacemaker.go
@@ -8,7 +8,6 @@ import (
 
 	consensusTelemetry "github.com/pokt-network/pocket/consensus/telemetry"
 	typesCons "github.com/pokt-network/pocket/consensus/types"
-
 	"github.com/pokt-network/pocket/shared/modules"
 )