+++ persistence/block.go
@@ -3,13 +3,13 @@ package persistence
 import (
 	"encoding/binary"
 	"encoding/hex"
-	"fmt"
 
 	"github.com/pokt-network/pocket/persistence/kvstore"
 	"github.com/pokt-network/pocket/persistence/types"
+	"github.com/pokt-network/pocket/shared/codec"
 )
 
-// OPTIMIZE(team): get from blockstore or keep in memory
+// OPTIMIZE: evaluate if it's faster to get this from the blockstore (or cache) than the SQL engine
 func (p PostgresContext) GetLatestBlockHeight() (latestHeight uint64, err error) {
 	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
@@ -20,8 +20,8 @@ func (p PostgresContext) GetLatestBlockHeight() (latestHeight uint64, err error)
 	return
 }
 
-// OPTIMIZE(team): get from blockstore or keep in cache/memory
-func (p PostgresContext) GetBlockHashAtHeight(height int64) ([]byte, error) {
+// OPTIMIZE: evaluate if it's  faster to get this from the blockstore (or cache) than the SQL engine
+func (p PostgresContext) GetBlockHash(height int64) ([]byte, error) {
 	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return nil, err
@@ -40,21 +40,6 @@ func (p PostgresContext) GetHeight() (int64, error) {
 	return p.Height, nil
 }
 
-func (p PostgresContext) GetPrevAppHash() (string, error) {
-	height, err := p.GetHeight()
-	if err != nil {
-		return "", err
-	}
-	if height <= 1 {
-		return "TODO: get from genesis", nil
-	}
-	block, err := p.blockstore.Get(heightToBytes(height - 1))
-	if err != nil {
-		return "", fmt.Errorf("error getting block hash for height %d even though it's in the database: %s", height, err)
-	}
-	return hex.EncodeToString(block), nil // TODO(#284): Return `block.Hash` instead of the hex encoded representation of the blockBz
-}
-
 func (p PostgresContext) TransactionExists(transactionHash string) (bool, error) {
 	hash, err := hex.DecodeString(transactionHash)
 	if err != nil {
@@ -71,64 +56,67 @@ func (p PostgresContext) TransactionExists(transactionHash string) (bool, error)
 	return true, err
 }
 
-func (p PostgresContext) indexTransactions() error {
-	// TODO: store in batch
-	for _, txResult := range p.GetTxResults() {
-		if err := p.txIndexer.Index(txResult); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 // DISCUSS: this might be retrieved from the block store - temporarily we will access it directly from the module
 //       following the pattern of the Consensus Module prior to pocket/issue-#315
 // TODO(#284): Remove blockProtoBytes from the interface
-func (p *PostgresContext) SetProposalBlock(blockHash string, blockProtoBytes, proposerAddr []byte, transactions [][]byte) error {
+func (p *PostgresContext) SetProposalBlock(blockHash string, proposerAddr, quorumCert []byte, transactions [][]byte) error {
 	p.blockHash = blockHash
-	p.blockProtoBytes = blockProtoBytes
+	p.quorumCert = quorumCert
 	p.proposerAddr = proposerAddr
 	p.blockTxs = transactions
 	return nil
 }
 
-// TEMPORARY: Including two functions for the SQL and KV Store as an interim solution
-//                 until we include the schema as part of the SQL Store because persistence
-//                 currently has no access to the protobuf schema which is the source of truth.
-// TODO: atomic operations needed here - inherited pattern from consensus module
-func (p PostgresContext) storeBlock(quorumCert []byte) error {
-	if p.blockProtoBytes == nil {
-		// IMPROVE/CLEANUP: HACK - currently tests call Commit() on the same height and it throws a
-		// ERROR: duplicate key value violates unique constraint "block_pkey", because it attempts to
-		// store a block at height 0 for each test. We need a cleanup function to clear the block table
-		// each iteration
-		return nil
+// Creates a block protobuf object using the schema defined in the persistence module
+func (p *PostgresContext) prepareBlock(quorumCert []byte) (*types.Block, error) {
+	var prevHash []byte
+	if p.Height == 0 {
+		prevHash = []byte("")
+	} else {
+		var err error
+		prevHash, err = p.GetBlockHash(p.Height - 1)
+		if err != nil {
+			return nil, err
+		}
 	}
-	// INVESTIGATE: Note that we are writing this directly to the blockStore. Depending on how
-	// the use of the PostgresContext evolves, we may need to write this to `ContextStore` and copy
-	// over to `BlockStore` when the block is committed.
-	if err := p.blockstore.Put(heightToBytes(p.Height), p.blockProtoBytes); err != nil {
-		return err
+
+	txsHash, err := p.getTxsHash()
+	if err != nil {
+		return nil, err
 	}
-	// Store in SQL Store
-	if err := p.InsertBlock(uint64(p.Height), p.blockHash, p.proposerAddr, quorumCert); err != nil {
-		return err
+
+	block := &types.Block{
+		Height:            uint64(p.Height),
+		StateHash:         p.blockHash,
+		PrevStateHash:     hex.EncodeToString(prevHash),
+		ProposerAddress:   p.proposerAddr,
+		QuorumCertificate: quorumCert,
+		TransactionsHash:  txsHash,
 	}
-	// Store transactions in indexer
-	return p.indexTransactions()
+
+	return block, nil
 }
 
-func (p PostgresContext) InsertBlock(height uint64, hash string, proposerAddr []byte, quorumCert []byte) error {
+// Inserts the block into the postgres database
+func (p *PostgresContext) insertBlock(block *types.Block) error {
 	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return err
 	}
 
-	_, err = tx.Exec(ctx, types.InsertBlockQuery(height, hash, proposerAddr, quorumCert))
+	_, err = tx.Exec(ctx, types.InsertBlockQuery(block.Height, block.StateHash, block.ProposerAddress, block.QuorumCertificate))
 	return err
 }
 
-// CLEANUP: Should this be moved to a shared directory?
+// Stores the block in the key-value store
+func (p PostgresContext) storeBlock(block *types.Block) error {
+	blockBz, err := codec.GetCodec().Marshal(block)
+	if err != nil {
+		return err
+	}
+	return p.blockStore.Set(heightToBytes(p.Height), blockBz)
+}
+
 func heightToBytes(height int64) []byte {
 	heightBytes := make([]byte, 8)
 	binary.LittleEndian.PutUint64(heightBytes, uint64(height))