+++ shared/messaging/envelope_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func Test_UnpackMessage_Roundtrip(t *testing.T) {
-	someMsg := &DebugMessage{Action: DebugMessageAction_DEBUG_CLEAR_STATE}
+	someMsg := &DebugMessage{Action: DebugMessageAction_DEBUG_PERSISTENCE_CLEAR_STATE}
 	packedMsg, err := PackMessage(someMsg)
 	require.NoError(t, err)