+++ shared/modules/types.go
@@ -47,6 +47,7 @@ type PersistenceConfig interface {
 	GetNodeSchema() string
 	GetBlockStorePath() string
 	GetTxIndexerPath() string
+	GetTreesStoreDir() string
 }
 
 type P2PConfig interface {