+++ shared/modules/doc/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.4] - 2022-11-30
+
+- Removed `GetPrevHash` and just using `GetBlockHash` instead
+- Removed `blockProtoBz` from `SetProposalBlock` interface
+- Removed `GetLatestBlockTxs` and `SetLatestTxResults` in exchange for `IndexTransaction`
+- Removed `SetTxResults`
+- Renamed `UpdateAppHash` to `ComputeStateHash`
+- Removed some getters related to the proposal block (`GetBlockTxs`, `GetBlockHash`, etc…)
+
 ## [0.0.0.3] - 2022-11-15
 
 PersistenceModule